+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules
+# Keep environment variables out of version control
+.env
