+++ back-end/package.json
@@ -20,7 +20,7 @@
     "eslint": "^8.48.0",
     "nodemon": "^3.0.1",
     "prettier": "^3.0.3",
-    "prisma": "^5.10.2"
+    "prisma": "^5.13.0"
   },
   "type": "module"
 }
