+++ back-end/src/routes/users.js
@@ -1,13 +1,13 @@
-import { Router } from 'express'
-import controller from '../controllers/users.js'
+// import { Router } from 'express'
+// import controller from '../controllers/users.js'
 
-const router = Router()
+// const router = Router()
 
-router.post('/', controller.create)
-router.get('/', controller.retrieveAll)
-router.get('/:id', controller.retrieveOne)
-router.put('/:id', controller.update)
-router.delete('/:id', controller.delete)
-router.post('/login', controller.login)
+// router.post('/', controller.create)
+// router.get('/', controller.retrieveAll)
+// router.get('/:id', controller.retrieveOne)
+// router.put('/:id', controller.update)
+// router.delete('/:id', controller.delete)
+// router.post('/login', controller.login)
 
-export default router
+// export default router
