+++ package.json
@@ -1,0 +1,8 @@
+{
+  "devDependencies": {
+    "prisma": "^5.13.0"
+  },
+  "dependencies": {
+    "@prisma/client": "^5.13.0"
+  }
+}
